+++ apps/web/app/(main)/[slug]/_components/product-description.tsx
@@ -6,15 +6,18 @@ export const ProductDescription: React.FC = () => {
   } = usePage()
 
   return (
-    <section className='rounded-lg bg-card p-6 shadow-md'>
-      <h3 className='mb-4 text-2xl font-semibold'>Description</h3>
+    <section className='rounded-lg bg-card p-6 shadow-md dark:border'>
+      <h2 className='mb-6 text-2xl font-semibold text-balance'>Overview</h2>
 
-      <p className='text-base text-pretty whitespace-pre-wrap'>
+      <p className='mb-8 text-base text-pretty whitespace-pre-wrap'>
         {description?.split('\\n').join('\n')}
       </p>
 
-      <h3 className='mt-6 mb-4 text-2xl font-semibold'>Attributes</h3>
-      <ul className='capitalize'>
+      <h3 className='mb-6 text-2xl font-semibold text-balance'>
+        Specifications
+      </h3>
+
+      <ul className='space-y-2 text-pretty capitalize'>
         {attributes.map((attr) => (
           <li key={attr.name} className='text-lg'>
             <strong>{attr.name}:</strong> {attr.value}