+++ apps/web/app/(main)/account/orders/page.tsx
@@ -1,3 +1,16 @@
+import { AccountHeader } from '@/app/(main)/account/_components/header'
+
 export default function AccountOrdersPage() {
-  return <div>Not implemented</div>
+  return (
+    <>
+      <AccountHeader
+        title='My Orders'
+        description='Review your past orders, track current shipments, and manage returns or exchanges all in one place.'
+      />
+
+      <section>
+        <h2 className='sr-only'>Orders History List section</h2>
+      </section>
+    </>
+  )
 }