+++ apps/web/app/(main)/account/wishlist/page.tsx
@@ -1,5 +1,6 @@
 import { Suspense } from 'react'
 
+import { AccountHeader } from '@/app/(main)/account/_components/header'
 import {
   WishlistItems,
   WishlistItemsSkeleton,
@@ -13,13 +14,18 @@ export default function AccountWishlistPage() {
 
   return (
     <HydrateClient>
-      <main className='container flex-1 py-6'>
-        <section className='grid grid-cols-1 gap-4 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4'>
-          <Suspense fallback={<WishlistItemsSkeleton />}>
-            <WishlistItems />
-          </Suspense>
-        </section>
-      </main>
+      <AccountHeader
+        title='My Wishlist'
+        description="View and manage the items you've saved to your wishlist for future reference or purchase."
+      />
+
+      <section className='grid max-h-full grid-cols-1 gap-4 overflow-y-auto px-6 pt-6 sm:grid-cols-2 lg:grid-cols-3'>
+        <h2 className='sr-only'>Wishlist Items List section</h2>
+
+        <Suspense fallback={<WishlistItemsSkeleton />}>
+          <WishlistItems />
+        </Suspense>
+      </section>
     </HydrateClient>
   )
 }