+++ apps/web/components/providers.tsx
@@ -3,24 +3,14 @@
 import type { QueryClient } from '@tanstack/react-query'
 import { useState } from 'react'
 import { QueryClientProvider } from '@tanstack/react-query'
-import {
-  createTRPCClient,
-  httpBatchLink,
-  httpBatchStreamLink,
-  retryLink,
-  splitLink,
-} from '@trpc/client'
 import { NuqsAdapter } from 'nuqs/adapters/next/app'
-import SuperJSON from 'superjson'
 
-import type { AppRouter } from '@yukinu/api'
 import { SessionProvider } from '@yukinu/auth/react'
 import { createQueryClient } from '@yukinu/lib/create-query-client'
 import { ThemeProvider } from '@yukinu/ui'
-import { env } from '@yukinu/validators/env.next'
 
+import { createClient } from '@/lib/trpc/create-client'
 import { TRPCProvider } from '@/lib/trpc/react'
-import { getWebUrl } from '@/lib/utils'
 
 let clientQueryClientSingleton: QueryClient | undefined = undefined
 const getQueryClient = () => {
@@ -33,39 +23,8 @@ export function Providers({
 }: Readonly<{ children: React.ReactNode }>) {
   const queryClient = getQueryClient()
 
-  const configs = {
-    transformer: SuperJSON,
-    url: getWebUrl() + '/api/trpc',
-    headers() {
-      const headers = new Headers()
-      headers.set('x-trpc-source', 'web')
-      return headers
-    },
-  }
-
   // eslint-disable-next-line @eslint-react/naming-convention/use-state
-  const [trpcClient] = useState(() =>
-    createTRPCClient<AppRouter>({
-      links: [
-        retryLink({
-          retry: ({ error, attempts }) => {
-            if (error.data?.code === 'UNAUTHORIZED' && attempts < 2) {
-              void fetch('/api/auth/refresh-token', { method: 'POST' })
-              return true
-            }
-
-            return false
-          },
-          retryDelayMs: (attempts) => Math.min(1000 * 2 ** attempts, 30000),
-        }),
-        splitLink({
-          condition: () => env.NEXT_PUBLIC_TRPC_USE_STREAMING === 'true',
-          true: httpBatchStreamLink(configs),
-          false: httpBatchLink(configs),
-        }),
-      ],
-    }),
-  )
+  const [trpcClient] = useState(createClient)
 
   return (
     <ThemeProvider attribute='class' disableTransitionOnChange enableSystem>