+++ apps/web/lib/trpc/rsc.tsx
@@ -6,6 +6,8 @@ import { createTRPCOptionsProxy } from '@trpc/tanstack-react-query'
 import { appRouter, createCaller, createTRPCContext } from '@yukinu/api'
 import { createQueryClient } from '@yukinu/lib/create-query-client'
 
+import { createClient } from '@/lib/trpc/create-client'
+
 const createRscContext = cache(async () => {
   const heads = new Headers(await headers())
   heads.set('x-trpc-source', 'rsc')
@@ -20,6 +22,7 @@ const api = createCaller(() => createRscContext())
 const trpc = createTRPCOptionsProxy({
   ctx: () => createRscContext(),
   queryClient: getQueryClient,
+  client: createClient(),
   router: appRouter,
 })